--- Assets/Materials/Paintable Character.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 945b2f97ece4627488fc022fb79f3396
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
