--- Assets/Materials/Paintable Wall 1.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 376ccc420567af148999de48914bc64c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
