--- Assets/Materials/Paintable Wall 2.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 82254c639e3a7e2409f83c6a2fd09e73
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
