--- Assets/Materials/Paintable Wall 3.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 51ab6d41e15891d469cbc70dc63ebe28
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
