--- Assets/Materials/Paintable Wall 4.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b6f8c5d6373dfdc49a0118c30785cb7b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
