--- Assets/Script/Game/Core/MousePainter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 54541a8f4c27a164a93ee8aaba3a3d95
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
