--- Assets/Script/Game/Core/Painter.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace PaintOnObject 
-{
-    public abstract class Painter : MonoBehaviour
-    {
-        public Color color;
-        [Min(0.01f)] public float radius = 1;
-        public float strength = 1;
-        public float hardness = 1;
-    }
-
-}
