+++ Assets/Script/Game/Core/PolygonPaintManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 79dc41b8a0467004892fb58be9a46fbe
+guid: cd7d4e8e7ce7d514e9a7c67277ced8e5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
