+++ Assets/Shaders/PolygonPainter.shader.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: 2ec6165a59449a642854dff8d828add4
+guid: 848d39d443743084990dd1b91b27ab6c
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
   nonModifiableTextures: []
-  preprocessorOverride: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
